+++ apps/demo-fuels/src/index.test.ts
@@ -5,25 +5,21 @@
  * It ensures that built code is fully working.
  */
 
-import { Provider, toHex, Wallet, FUEL_NETWORK_URL, FuelError, ErrorCode } from 'fuels';
-import { expectToThrowFuelError, generateTestWallet } from 'fuels/test-utils';
+import { toHex, Wallet } from 'fuels';
+import { launchTestNode } from 'fuels/test-utils';
 
 import { SampleFactory, Sample } from './sway-programs-api';
 
-let baseAssetId: string;
-
 /**
  * @group node
+ * @group browser
  */
 describe('ExampleContract', () => {
-  beforeAll(async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
-    baseAssetId = provider.getBaseAssetId();
-  });
-
   it('should return the input', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
-    const wallet = await generateTestWallet(provider, [[500_000, baseAssetId]]);
+    using launched = await launchTestNode();
+    const {
+      wallets: [wallet],
+    } = launched;
 
     // Deploy
     const deploy = await SampleFactory.deploy(wallet);
@@ -44,8 +40,10 @@ describe('ExampleContract', () => {
   });
 
   it('should deploy contract', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
-    const wallet = await generateTestWallet(provider, [[500_000, baseAssetId]]);
+    using launched = await launchTestNode();
+    const {
+      wallets: [wallet],
+    } = launched;
 
     // Deploy
     const deploy = await SampleFactory.deploy(wallet);
@@ -60,8 +58,12 @@ describe('ExampleContract', () => {
   });
 
   it('should throw when simulating via contract factory with wallet with no resources', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
-    const fundedWallet = await generateTestWallet(provider, [[500_000, baseAssetId]]);
+    using launched = await launchTestNode();
+    const {
+      provider,
+      wallets: [fundedWallet],
+    } = launched;
+
     const unfundedWallet = Wallet.generate({ provider });
 
     const deploy = await SampleFactory.deploy(fundedWallet);
@@ -79,8 +81,12 @@ describe('ExampleContract', () => {
   });
 
   it('should not throw when dry running via contract factory with wallet with no resources', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
-    const fundedWallet = await generateTestWallet(provider, [[500_000, baseAssetId]]);
+    using launched = await launchTestNode();
+    const {
+      provider,
+      wallets: [fundedWallet],
+    } = launched;
+
     const unfundedWallet = Wallet.generate({ provider });
 
     const { waitForResult } = await SampleFactory.deploy(fundedWallet);
@@ -91,12 +97,16 @@ describe('ExampleContract', () => {
   });
 
   it('should demo how to use generated files just fine', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
-    const wallet = await generateTestWallet(provider, [[500_000, baseAssetId]]);
+    using launched = await launchTestNode();
+    const {
+      wallets: [wallet],
+    } = launched;
+
     const deploy = await SampleFactory.deploy(wallet);
-    const { contract: depoloyed } = await deploy.waitForResult();
+    const { contract: deployed } = await deploy.waitForResult();
+
     const contractsIds = {
-      sample: depoloyed.id,
+      sample: deployed.id,
     };
 
     // #region using-generated-files