+++ apps/docs-snippets/src/guide/scripts/script-with-configurable.test.ts
@@ -1,28 +1,21 @@
-import type { WalletUnlocked } from 'fuels';
 import { Script, BN } from 'fuels';
+import { launchTestNode } from 'fuels/test-utils';
 
-import {
-  DocSnippetProjectsEnum,
-  getDocsSnippetsForcProject,
-} from '../../../test/fixtures/forc-projects';
-import { getTestWallet } from '../../utils';
+import { SumScript } from '../../../test/typegen';
 
 /**
  * @group node
+ * @group browser
  */
-describe(__filename, () => {
-  let wallet: WalletUnlocked;
-  const { abiContents, binHexlified } = getDocsSnippetsForcProject(
-    DocSnippetProjectsEnum.SUM_SCRIPT
-  );
+describe('Script With Configurable', () => {
+  it('should successfully sum set configurable constants with inputted value', async () => {
+    using launched = await launchTestNode();
+    const {
+      wallets: [wallet],
+    } = launched;
 
-  beforeAll(async () => {
-    wallet = await getTestWallet();
-  });
-
-  it('should successfully sum setted configurable constant with inputted value', async () => {
     // #region script-with-configurable-contants-2
-    const script = new Script(binHexlified, abiContents, wallet);
+    const script = new Script(SumScript.bytecode, SumScript.abi, wallet);
 
     const configurableConstants = {
       AMOUNT: 81,
@@ -42,11 +35,16 @@ describe(__filename, () => {
   });
 
   it('prepares a script and retrieves the id before submission', async () => {
+    using launched = await launchTestNode();
+    const {
+      wallets: [wallet],
+    } = launched;
+
     const argument = 10;
     const expected = 20;
 
     // #region preparing-scripts
-    const script = new Script(binHexlified, abiContents, wallet);
+    const script = new Script(SumScript.bytecode, SumScript.abi, wallet);
 
     const tx = script.functions.main(argument);