+++ internal/fuel-core/lib/shared.js
@@ -51,23 +51,23 @@ const fuelCoreRepoUrl = 'https://github.com/fuellabs/fuel-core.git';
 
 export const buildFromGitBranch = (branchName) => {
   const fuelCoreRepoDir = join(__dirname, '..', 'fuel-core-repo');
-  const fuelCoreRepoDebugDir = join(fuelCoreRepoDir, 'target', 'debug');
+  const fuelCoreRepoReleaseDir = join(fuelCoreRepoDir, 'target', 'release');
   const stdioOpts = { stdio: 'inherit' };
 
   if (existsSync(fuelCoreRepoDir)) {
-    spawnSync('git', ['pull'], { cwd: fuelCoreRepoDir, ...stdioOpts });
     spawnSync('git', ['checkout', branchName], { cwd: fuelCoreRepoDir, ...stdioOpts });
-    spawnSync('cargo', ['build'], { cwd: fuelCoreRepoDir, ...stdioOpts });
+    spawnSync('git', ['pull'], { cwd: fuelCoreRepoDir, ...stdioOpts });
+    spawnSync('cargo', ['build', '--release'], { cwd: fuelCoreRepoDir, ...stdioOpts });
   } else {
     spawnSync(
       'git',
       ['clone', '--branch', branchName, fuelCoreRepoUrl, fuelCoreRepoDir],
       stdioOpts
     );
-    spawnSync('cargo', ['build'], { cwd: fuelCoreRepoDir, ...stdioOpts });
+    spawnSync('cargo', ['build', '--release'], { cwd: fuelCoreRepoDir, ...stdioOpts });
   }
 
-  const [from, to] = [fuelCoreRepoDebugDir, fuelCoreBinDirPath];
+  const [from, to] = [fuelCoreRepoReleaseDir, fuelCoreBinDirPath];
 
   rmSync(to, { recursive: true, force: true });
   mkdirSync(to, { recursive: true });