+++ packages/account/src/providers/provider.test.ts
@@ -16,13 +16,7 @@ import {
   MESSAGE_PROOF_RAW_RESPONSE,
   MESSAGE_PROOF,
 } from '../../test/fixtures';
-import {
-  setupTestProviderAndWallets,
-  launchNode,
-  seedTestWallet,
-  TestMessage,
-} from '../test-utils';
-import { Wallet } from '../wallet';
+import { setupTestProviderAndWallets, launchNode, TestMessage } from '../test-utils';
 
 import type { Coin } from './coin';
 import { coinQuantityfy } from './coin-quantity';
@@ -41,10 +35,6 @@ import { TransactionResponse } from './transaction-response';
 import type { SubmittedStatus } from './transaction-summary/types';
 import * as gasMod from './utils/gas';
 
-afterEach(() => {
-  vi.restoreAllMocks();
-});
-
 const getCustomFetch =
   (expectedOperationName: string, expectedResponse: object) =>
   async (url: string, options: RequestInit | undefined) => {
@@ -62,9 +52,6 @@ const getCustomFetch =
     return fetch(url, options);
   };
 
-// TODO: Figure out a way to import this constant from `@fuel-ts/account/configs`
-const FUEL_NETWORK_URL = 'http://127.0.0.1:4000/v1/graphql';
-
 /**
  * @group node
  */
@@ -75,7 +62,7 @@ describe('Provider', () => {
 
     const version = await provider.getVersion();
 
-    expect(version).toEqual('0.32.1');
+    expect(version).toEqual('0.33.0');
   });
 
   it('can call()', async () => {
@@ -231,14 +218,18 @@ describe('Provider', () => {
   });
 
   it('gets the chain ID', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
+
     const chainId = provider.getChainId();
 
     expect(chainId).toBe(0);
   });
 
   it('gets the base asset ID', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
+
     const baseAssetId = provider.getBaseAssetId();
 
     expect(baseAssetId).toBeDefined();
@@ -753,7 +744,10 @@ describe('Provider', () => {
   it('can getMessageProof with all data', async () => {
     // Create a mock provider to return the message proof
     // It test mainly types and converstions
-    const provider = await Provider.create(FUEL_NETWORK_URL, {
+    using launched = await setupTestProviderAndWallets();
+    const { provider: nodeProvider } = launched;
+
+    const provider = await Provider.create(nodeProvider.url, {
       fetch: async (url, options) =>
         getCustomFetch('getMessageProof', { messageProof: MESSAGE_PROOF_RAW_RESPONSE })(
           url,
@@ -773,7 +767,10 @@ describe('Provider', () => {
   it('can getMessageStatus', async () => {
     // Create a mock provider to return the message proof
     // It test mainly types and converstions
-    const provider = await Provider.create(FUEL_NETWORK_URL, {
+    using launched = await setupTestProviderAndWallets();
+    const { provider: nodeProvider } = launched;
+
+    const provider = await Provider.create(nodeProvider.url, {
       fetch: async (url, options) =>
         getCustomFetch('getMessageStatus', { messageStatus: messageStatusResponse })(url, options),
     });
@@ -903,7 +900,10 @@ describe('Provider', () => {
 
     const consoleWarnSpy = vi.spyOn(console, 'warn');
 
-    await Provider.create(FUEL_NETWORK_URL);
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
+
+    await Provider.create(provider.url);
 
     expect(consoleWarnSpy).toHaveBeenCalledOnce();
     expect(consoleWarnSpy).toHaveBeenCalledWith(
@@ -935,7 +935,10 @@ Supported fuel-core version: ${mock.supportedVersion}.`
 
     const consoleWarnSpy = vi.spyOn(console, 'warn');
 
-    await Provider.create(FUEL_NETWORK_URL);
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
+
+    await Provider.create(provider.url);
 
     expect(consoleWarnSpy).toHaveBeenCalledOnce();
     expect(consoleWarnSpy).toHaveBeenCalledWith(
@@ -946,13 +949,57 @@ Supported fuel-core version: ${mock.supportedVersion}.`
     );
   });
 
+  it('should ensure fuel node version warning is shown before chain incompatibility error', async () => {
+    const { FUEL_CORE } = versions;
+    const [major, minor, patch] = FUEL_CORE.split('.');
+    const majorMismatch = major === '0' ? 1 : parseInt(patch, 10) - 1;
+
+    const mock = {
+      isMajorSupported: false,
+      isMinorSupported: true,
+      isPatchSupported: true,
+      supportedVersion: `${majorMismatch}.${minor}.${patch}`,
+    };
+
+    if (mock.supportedVersion === FUEL_CORE) {
+      throw new Error();
+    }
+
+    const spy = vi.spyOn(fuelTsVersionsMod, 'checkFuelCoreVersionCompatibility');
+    spy.mockImplementationOnce(() => mock);
+
+    const consoleWarnSpy = vi.spyOn(console, 'warn');
+
+    const graphQLDummyError = `Unknown field "height" on type "Block".
+      Unknown field "version" on type "ScriptParameters".
+      Unknown field "version" on type "ConsensusParameters".`;
+
+    const fuelError = new FuelError(ErrorCode.INVALID_REQUEST, graphQLDummyError);
+
+    const fetchChainSpy = vi
+      .spyOn(Provider.prototype, 'fetchChain')
+      .mockImplementationOnce(async () => Promise.reject(fuelError));
+
+    await expectToThrowFuelError(() => setupTestProviderAndWallets(), fuelError);
+
+    expect(consoleWarnSpy).toHaveBeenCalledOnce();
+    expect(consoleWarnSpy).toHaveBeenCalledWith(
+      `The Fuel Node that you are trying to connect to is using fuel-core version ${FUEL_CORE},
+which is not supported by the version of the TS SDK that you are using.
+Things may not work as expected.
+Supported fuel-core version: ${mock.supportedVersion}.`
+    );
+
+    expect(fetchChainSpy).toHaveBeenCalledOnce();
+  });
+
   it('An invalid subscription request throws a FuelError and does not hold the test runner (closes all handles)', async () => {
     using launched = await setupTestProviderAndWallets();
     const { provider } = launched;
 
     await expectToThrowFuelError(
       async () => {
-        for await (const value of provider.operations.statusChange({
+        for await (const value of await provider.operations.statusChange({
           transactionId: 'invalid transaction id',
         })) {
           // shouldn't be reached and should fail if reached
@@ -1009,7 +1056,7 @@ Supported fuel-core version: ${mock.supportedVersion}.`
     );
 
     const { error } = await safeExec(async () => {
-      for await (const iterator of provider.operations.statusChange({
+      for await (const iterator of await provider.operations.statusChange({
         transactionId: 'doesnt matter, will be aborted',
       })) {
         // shouldn't be reached and should fail if reached
@@ -1123,7 +1170,8 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   it('subscriptions: does not throw when stream contains more than one "data:"', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
 
     vi.spyOn(global, 'fetch').mockImplementationOnce(() => {
       const responseObject = {
@@ -1149,7 +1197,7 @@ Supported fuel-core version: ${mock.supportedVersion}.`
       );
     });
 
-    for await (const { submitAndAwait } of provider.operations.submitAndAwait({
+    for await (const { submitAndAwait } of await provider.operations.submitAndAwait({
       encodedTransaction: "it's mocked so doesn't matter",
     })) {
       expect(submitAndAwait.type).toEqual('SuccessStatus');
@@ -1158,7 +1206,8 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   test('subscriptions: ignores keep-alive messages', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
 
     const fetchSpy = vi.spyOn(global, 'fetch');
 
@@ -1180,7 +1229,7 @@ Supported fuel-core version: ${mock.supportedVersion}.`
 
     let numberOfEvents = 0;
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    for await (const { submitAndAwait } of provider.operations.submitAndAwait({
+    for await (const { submitAndAwait } of await provider.operations.submitAndAwait({
       encodedTransaction: "it's mocked so doesn't matter",
     })) {
       numberOfEvents += 1;
@@ -1190,7 +1239,8 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   it('subscriptions: does not throw when stream has two events in the same chunk', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
 
     vi.spyOn(global, 'fetch').mockImplementationOnce(() => {
       const event1 = {
@@ -1226,7 +1276,7 @@ Supported fuel-core version: ${mock.supportedVersion}.`
 
     let numberOfEvents = 0;
 
-    for await (const { submitAndAwait } of provider.operations.submitAndAwait({
+    for await (const { submitAndAwait } of await provider.operations.submitAndAwait({
       encodedTransaction: "it's mocked so doesn't matter",
     })) {
       numberOfEvents += 1;
@@ -1242,7 +1292,8 @@ Supported fuel-core version: ${mock.supportedVersion}.`
     expect(numberOfEvents).toEqual(2);
   });
   it('subscriptions: does not throw when an event is streamed in multiple chunks', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
 
     vi.spyOn(global, 'fetch').mockImplementationOnce(() => {
       const responseObject = JSON.stringify({
@@ -1273,7 +1324,7 @@ Supported fuel-core version: ${mock.supportedVersion}.`
       );
     });
 
-    for await (const { submitAndAwait } of provider.operations.submitAndAwait({
+    for await (const { submitAndAwait } of await provider.operations.submitAndAwait({
       encodedTransaction: "it's mocked so doesn't matter",
     })) {
       expect(submitAndAwait.type).toEqual('SuccessStatus');
@@ -1281,7 +1332,8 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   it('subscriptions: does not throw when chunk has a full and partial event in it', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
 
     vi.spyOn(global, 'fetch').mockImplementationOnce(() => {
       const event1 = {
@@ -1320,7 +1372,7 @@ Supported fuel-core version: ${mock.supportedVersion}.`
 
     let numberOfEvents = 0;
 
-    for await (const { submitAndAwait } of provider.operations.submitAndAwait({
+    for await (const { submitAndAwait } of await provider.operations.submitAndAwait({
       encodedTransaction: "it's mocked so doesn't matter",
     })) {
       numberOfEvents += 1;
@@ -1337,7 +1389,8 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   it('subscriptions: does not throw when multiple chunks contain multiple events with a keep-alive message in-between', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
 
     vi.spyOn(global, 'fetch').mockImplementationOnce(() => {
       const event1 = JSON.stringify({
@@ -1378,7 +1431,7 @@ Supported fuel-core version: ${mock.supportedVersion}.`
 
     let numberOfEvents = 0;
 
-    for await (const { submitAndAwait } of provider.operations.submitAndAwait({
+    for await (const { submitAndAwait } of await provider.operations.submitAndAwait({
       encodedTransaction: "it's mocked so doesn't matter",
     })) {
       numberOfEvents += 1;
@@ -1395,7 +1448,8 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   it('subscriptions: throws if the stream data string parsing fails for some reason', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
 
     const badResponse = 'data: {f: {}\n\n';
     vi.spyOn(global, 'fetch').mockImplementationOnce(() => {
@@ -1415,7 +1469,7 @@ Supported fuel-core version: ${mock.supportedVersion}.`
     await expectToThrowFuelError(
       async () => {
         // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        for await (const { submitAndAwait } of provider.operations.submitAndAwait({
+        for await (const { submitAndAwait } of await provider.operations.submitAndAwait({
           encodedTransaction: "it's mocked so doesn't matter",
         })) {
           // shouldn't be reached!
@@ -1429,8 +1483,11 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   test('requestMiddleware modifies the request before being sent to the node [sync]', async () => {
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
+
     const fetchSpy = vi.spyOn(global, 'fetch');
-    await Provider.create(FUEL_NETWORK_URL, {
+    await Provider.create(provider.url, {
       requestMiddleware: (request) => {
         request.headers ??= {};
         (request.headers as Record<string, string>)['x-custom-header'] = 'custom-value';
@@ -1446,8 +1503,11 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   test('requestMiddleware modifies the request before being sent to the node [async]', async () => {
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
+
     const fetchSpy = vi.spyOn(global, 'fetch');
-    await Provider.create(FUEL_NETWORK_URL, {
+    await Provider.create(provider.url, {
       requestMiddleware: (request) => {
         request.headers ??= {};
         (request.headers as Record<string, string>)['x-custom-header'] = 'custom-value';
@@ -1463,8 +1523,11 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   test('requestMiddleware works for subscriptions', async () => {
+    using launched = await setupTestProviderAndWallets();
+    const { provider: nodeProvider } = launched;
+
     const fetchSpy = vi.spyOn(global, 'fetch');
-    const provider = await Provider.create(FUEL_NETWORK_URL, {
+    const provider = await Provider.create(nodeProvider.url, {
       requestMiddleware: (request) => {
         request.headers ??= {};
         (request.headers as Record<string, string>)['x-custom-header'] = 'custom-value';
@@ -1473,7 +1536,7 @@ Supported fuel-core version: ${mock.supportedVersion}.`
     });
 
     await safeExec(async () => {
-      for await (const iterator of provider.operations.statusChange({
+      for await (const iterator of await provider.operations.statusChange({
         transactionId: 'doesnt matter, will be aborted',
       })) {
         // Just running a subscription to trigger the middleware
@@ -1491,8 +1554,11 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   test('custom fetch works with requestMiddleware', async () => {
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
+
     let requestHeaders: HeadersInit | undefined;
-    await Provider.create(FUEL_NETWORK_URL, {
+    await Provider.create(provider.url, {
       fetch: async (url, requestInit) => {
         requestHeaders = requestInit?.headers;
         return fetch(url, requestInit);
@@ -1510,8 +1576,11 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   test('custom fetch works with timeout', async () => {
+    using launched = await setupTestProviderAndWallets();
+    const { provider: nodeProvider } = launched;
+
     const timeout = 500;
-    const provider = await Provider.create(FUEL_NETWORK_URL, {
+    const provider = await Provider.create(nodeProvider.url, {
       fetch: async (url, requestInit) => fetch(url, requestInit),
       timeout,
     });
@@ -1533,7 +1602,8 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   test('getMessageByNonce', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
+    using launched = await setupTestProviderAndWallets();
+    const { provider } = launched;
 
     const nonce = '0x381de90750098776c71544527fd253412908dec3d07ce9a7367bd1ba975908a0';
     const message = await provider.getMessageByNonce(nonce);
@@ -1791,23 +1861,27 @@ Supported fuel-core version: ${mock.supportedVersion}.`
   });
 
   test('can properly use getBalances', async () => {
-    const provider = await Provider.create(FUEL_NETWORK_URL);
-    const baseAssetId = provider.getBaseAssetId();
-    const wallet = Wallet.generate({ provider });
-
     const fundAmount = 10_000;
 
-    await seedTestWallet(wallet, [
-      [fundAmount, baseAssetId],
-      [fundAmount, ASSET_A],
-    ]);
+    using launched = await setupTestProviderAndWallets({
+      walletsConfig: {
+        amountPerCoin: fundAmount,
+      },
+    });
+    const {
+      provider,
+      wallets: [wallet],
+    } = launched;
+
+    const baseAssetId = provider.getBaseAssetId();
 
     const { balances } = await provider.getBalances(wallet.address);
 
-    expect(balances.length).toBe(2);
+    expect(balances.length).toBe(3);
+
     balances.forEach((balance) => {
       expect(balance.amount.toNumber()).toBe(fundAmount);
-      expect([baseAssetId, ASSET_A].includes(balance.assetId)).toBeTruthy();
+      expect([baseAssetId, ASSET_A, ASSET_B].includes(balance.assetId)).toBeTruthy();
     });
   });
 });