+++ packages/account/src/providers/transaction-summary/types.ts
@@ -136,6 +136,7 @@ export type InputOutputParam = InputParam & OutputParam;
 
 export interface GetTransferOperationsParams extends InputOutputParam {
   receipts: TransactionResultReceipt[];
+  baseAssetId: string;
 }
 
 export type GetOperationParams = {
@@ -157,7 +158,7 @@ export interface MintedAsset {
 export type BurnedAsset = MintedAsset;
 
 export type TransactionSummary<TTransactionType = void> = {
-  id?: string;
+  id: string;
   time?: string;
   operations: Operation[];
   gasUsed: BN;