+++ packages/account/src/test-utils.ts
@@ -1,5 +1,3 @@
-export * from './test-utils/generateTestWallet';
-export * from './test-utils/seedTestWallet';
 export * from './test-utils/launchNode';
 export * from './test-utils/setup-test-provider-and-wallets';
 export * from './test-utils/wallet-config';